refactor: layout folder
