--- src/layouts/app/header-with-asidebar/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./_HeaderWithAsideBarLayout"
