--- src/layouts/app/header-with-asidebar/job-detail-header/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./__JobDetailHeaderLayout"
