--- src/layouts/app/header-with-asidebar/job-detail-header/job-candidate-info/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./___JobCandidateInfoLayout"
