+++ src/layouts/app/index.ts
@@ -1,1 +1,3 @@
-export { default } from "./AppLayout"
+export { default as AppHeaderWithAsideBarLayout } from "./_app-header-with-aside-bar"
+export { default as JobDetailLayout } from "./__job-detail-header"
+export { default as JobDetailCandidateInfoLayout } from "./___job-detail-candidate-info"
